+++ ChangeLog.md
@@ -3,6 +3,9 @@
 ## [Unrelease]
 ### [Added]
 - Add `WebMercator` to convert coordinate.
+- Add class `myMap` to get a map from OSM, and then implemented methods:
+	- constructor
+	- `getTileCoord()`
 
 ### [Changed]